Use usage-advanced as only usage file
